#28 address review: colocate components, use react hook form, fix view link, clean up actions
